requirements.txt
streamlit>=1.28.0
numpy>=1.24.0
pypdf2>=3.0.0
python-docx>=1.1.0
requests>=2.31.0
langchain==0.1.0
faiss-cpu==1.7.4
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.2
accelerate==0.24.1
huggingface-hub==0.19.4